enable Modbus Host mode for RTU implementation again, resolve #40
